+++ Page/Component/Header/Header.php
@@ -1,16 +1,13 @@
 <?php
+
 /**
  * Copyright © OXID eSales AG. All rights reserved.
  * See LICENSE file for license details.
  */
 
-namespace OxidEsales\Codeception\Page\Component\Header;
+declare(strict_types=1);
 
-use OxidEsales\Codeception\Page\Component\Header\AccountMenu;
-use OxidEsales\Codeception\Page\Component\Header\LanguageMenu;
-use OxidEsales\Codeception\Page\Component\Header\MiniBasket;
-use OxidEsales\Codeception\Page\Component\Header\Navigation;
-use OxidEsales\Codeception\Page\Component\Header\SearchWidget;
+namespace OxidEsales\Codeception\Page\Component\Header;
 
 /**
  * Trait for account menu widget.