+++ CHANGELOG.md
@@ -4,10 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [4.0.1] - 2022-11-23
+## [4.1.0] - unreleased
 
 ### Fixed
 - Checks for Page objects' availability
+### Added
+- Page objects:
+  - `Page\Lists\DistributorList`
+  - `Page\Lists\ManufacturerList`
+- Traits:
+  - `Page\Component\Footer\CategoryWidget`
+  - `Page\Component\Footer\InformationWidget`
+  - `Page\Component\Footer\ManufacturerWidget`
+  - `Page\Component\Footer\Footer`
+  - `Page\Component\Header\Header`
+- Methods:
+  - `Page\Component\Modal::closeModalBox()`
+### Deprecated
+- `Page\Lists\ProductList::openDetailsPage()`
 
 ## [4.0.0] - 2022-10-28
 
@@ -251,7 +265,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - First version of the module introduced
 
-[4.0.1]: https://github.com/OXID-eSales/codeception-page-object/compare/v4.0.0..v4.0.1
 [4.0.0]: https://github.com/OXID-eSales/codeception-page-object/compare/v3.0.0..v4.0.0
 [3.0.0]: https://github.com/OXID-eSales/codeception-page-object/compare/v2.3.0..v3.0.0
 [2.3.0]: https://github.com/OXID-eSales/codeception-page-object/compare/v2.2.0..v2.3.0